+++ src/test/java/ninja/rail/suits/PassengerFormTest.java
@@ -1,4 +1,4 @@
-package ninja.rail.tests;
+package ninja.rail.suits;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.Epic;
